agri-agent-system/requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-google-genai>=2.0.0
pydantic>=2.0.0

trafilatura>=1.6.0
beautifulsoup4>=4.12.0
duckduckgo-search>=6.2.0
charset-normalizer>=3.0.0

jupyterlab>=4.0.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.13.0

streamlit>=1.30.0
python-dotenv>=1.0.0

chromadb>=0.5.0
tavily-python>=0.3.0

